+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New in v10.4
 --------------------------
+- Auto create main formula if not exists when you import a formula to make data
 - For full details please refer to the CHANGELOG